+++ package.json
@@ -21,7 +21,7 @@
     "web": "scripts/set-pusher-suffix.sh && concurrently npm:web-proxy npm:web-server",
     "web-proxy": "ts-node web/proxy.ts",
     "web-server": "webpack-dev-server --open --config config/webpack/webpack.dev.ts",
-    "build": "webpack --config config/webpack/webpack.common.ts --env file=.env.production",
+    "build": "webpack --config config/webpack/webpack.common.ts --env file=.env.production && ts-node scripts/combine-web-sourcemaps.ts",
     "build-staging": "webpack --config config/webpack/webpack.common.ts --env file=.env.staging",
     "build-adhoc": "webpack --config config/webpack/webpack.common.ts --env file=.env.adhoc",
     "desktop": "scripts/set-pusher-suffix.sh && ts-node desktop/start.ts",
@@ -53,6 +53,7 @@
     "symbolicate:ios": "npx metro-symbolicate main.jsbundle.map",
     "symbolicate-release:ios": "scripts/release-profile.ts --platform=ios",
     "symbolicate-release:android": "scripts/release-profile.ts --platform=android",
+    "symbolicate-release:web": "scripts/release-profile.ts --platform=web",
     "symbolicate-profile": "scripts/symbolicate-profile.ts",
     "combine-web-sourcemaps": "scripts/combine-web-sourcemaps.ts",
     "test:e2e": "ts-node tests/e2e/testRunner.ts --config ./config.local.ts",
@@ -262,7 +263,6 @@
     "babel-plugin-transform-class-properties": "^6.24.1",
     "babel-plugin-transform-remove-console": "^6.9.4",
     "clean-webpack-plugin": "^4.0.0",
-    "concat-with-sourcemaps": "^1.1.0",
     "concurrently": "^8.2.2",
     "copy-webpack-plugin": "^10.1.0",
     "css-loader": "^6.7.2",
@@ -306,7 +306,6 @@
     "source-map": "^0.7.4",
     "storybook": "^8.1.10",
     "style-loader": "^2.0.0",
-    "terser-webpack-plugin": "^5.3.10",
     "time-analytics-webpack-plugin": "^0.1.17",
     "ts-jest": "^29.1.2",
     "ts-node": "^10.9.2",
